Add initial R scripts, config.mk and Makefile
R scripts are simplified re-write of previous get demographic_data.R
scripts. Retrieves all demographic data for ARTR HECO POSE and PSSP.

Add makefile that call R scripts that get vital rate data, cleans temp
data and creates zip archive of project. The zip archive can be used
for testing that all the folders and data that need to be included
are included in the makefile.