Modify scripts to produce stan datalists
Adding scripts to compile datalists for each species and vital rate.
Updating makefile.  make all now produces stan datalists for analysis.
Renaming some files so that the names are more consistent.

Biggest change is that the stan datalists are now assembled by species
and vital rate separately. This means that the dependencies in the
makefile pipeline should be as limited as possible for the datalist.